+++ madgraph/iolibs/template_files/madmatrix/timermap.h
@@ -1,6 +1,6 @@
 // Copyright (C) 2020-2026 CERN and UCLouvain.
 // Licensed under the GNU Lesser General Public License (version 3 or later).
-// Created originally by: A. Valassi (Jul 2020) for the MG5aMC CUDACPP plugin.
+// Created originally by: A. Valassi (Jul 2020) for the MadGraph7 CUDACPP plugin.
 // Further modified by: O. Mattelaer, S. Roiser, A. Valassi (2020-2024).
 // Integrated with the MadGraph7 project in Feb 2026.