+++ tests/IOTests.py
@@ -1,12 +1,12 @@
 ################################################################################
 #
-# Copyright (c) 2009 The MadGraph5_aMC@NLO Development team and Contributors
+# Copyright (c) 2009 The MadGraph7 Development team and Contributors
 #
-# This file is a part of the MadGraph5_aMC@NLO project, an application which 
+# This file is a part of the MadGraph7 project, an application which 
 # automatically generates Feynman diagrams and matrix elements for arbitrary
 # high-energy processes in the Standard Model and beyond.
 #
-# It is subject to the MadGraph5_aMC@NLO license which should accompany this 
+# It is subject to the MadGraph7 license which should accompany this 
 # distribution.
 #
 # For more information, visit madgraph.phys.ucl.ac.be and amcatnlo.web.cern.ch
@@ -778,9 +778,9 @@ class IOTestManager(unittest.TestCase):
                     # W(I)%P) are left alone — no need to escape them as '%%'.
                     # So that the version and date is automatically updated
                     target = target.replace(
-                        misc.apply_template('MadGraph5_aMC@NLO v. %(version)s, %(date)s',
+                        misc.apply_template('MadGraph7 v. %(version)s, %(date)s',
                                             misc.get_pkg_info()),
-                        'MadGraph5_aMC@NLO v. %(version)s, %(date)s')
+                        'MadGraph7 v. %(version)s, %(date)s')
                     target = target.replace(
                         misc.apply_template('v%(version)s (%(date)s)',
                                             misc.get_pkg_info()),